Extracted automatic context propagation tests to another class
