+++ reactor-core/src/withMicrometerTest/java/reactor/core/publisher/ContextPropagationTest.java
@@ -16,22 +16,10 @@
 
 package reactor.core.publisher;
 
-import java.time.Duration;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Queue;
-import java.util.concurrent.ArrayBlockingQueue;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CompletionStage;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.BiConsumer;
 import java.util.function.Function;
 
@@ -41,7 +29,6 @@ import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.provider.EnumSource;
@@ -57,17 +44,13 @@ import reactor.core.publisher.FluxHandle.HandleConditionalSubscriber;
 import reactor.core.publisher.FluxHandle.HandleSubscriber;
 import reactor.core.publisher.FluxHandleFuseable.HandleFuseableConditionalSubscriber;
 import reactor.core.publisher.FluxHandleFuseable.HandleFuseableSubscriber;
-import reactor.core.scheduler.Schedulers;
 import reactor.test.ParameterizedTestWithName;
-import reactor.test.publisher.TestPublisher;
 import reactor.test.subscriber.TestSubscriber;
 import reactor.test.subscriber.TestSubscriberBuilder;
-import reactor.util.concurrent.Queues;
 import reactor.util.context.Context;
 import reactor.util.context.ContextView;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
 /**
  * @author Simon Baslé
@@ -106,226 +89,6 @@ class ContextPropagationTest {
 	}
 
 	@Test
-	void threadLocalsPresentAfterSubscribeOn() {
-		Hooks.enableAutomaticContextPropagation();
-
-		AtomicReference<String> tlValue = new AtomicReference<>();
-
-		Flux.just(1)
-		    .subscribeOn(Schedulers.boundedElastic())
-		    .doOnNext(i -> tlValue.set(REF1.get()))
-		    .contextWrite(Context.of(KEY1, "present"))
-		    .blockLast();
-
-		assertThat(tlValue.get()).isEqualTo("present");
-	}
-
-	@Test
-	void threadLocalsPresentAfterPublishOn() {
-		Hooks.enableAutomaticContextPropagation();
-
-		AtomicReference<String> tlValue = new AtomicReference<>();
-
-		Flux.just(1)
-		    .publishOn(Schedulers.boundedElastic())
-		    .doOnNext(i -> tlValue.set(REF1.get()))
-		    .contextWrite(Context.of(KEY1, "present"))
-		    .blockLast();
-
-		assertThat(tlValue.get()).isEqualTo("present");
-	}
-
-	@Test
-	void threadLocalsPresentInFlatMap() {
-		Hooks.enableAutomaticContextPropagation();
-
-		AtomicReference<String> tlValue = new AtomicReference<>();
-
-		Flux.just(1)
-		    .flatMap(i -> Mono.just(i)
-		                      .doOnNext(j -> tlValue.set(REF1.get())))
-		    .contextWrite(Context.of(KEY1, "present"))
-		    .blockLast();
-
-		assertThat(tlValue.get()).isEqualTo("present");
-	}
-
-	@Test
-	void threadLocalsPresentAfterDelay() {
-		Hooks.enableAutomaticContextPropagation();
-
-		AtomicReference<String> tlValue = new AtomicReference<>();
-
-		Flux.just(1)
-		    .delayElements(Duration.ofMillis(1))
-		    .doOnNext(i -> tlValue.set(REF1.get()))
-		    .contextWrite(Context.of(KEY1, "present"))
-		    .blockLast();
-
-		assertThat(tlValue.get()).isEqualTo("present");
-	}
-
-	@Test
-	@Disabled("queue wrapping is removed starting from 3.5.7")
-	void threadLocalsRestoredAfterPollution() {
-		// this test validates Queue wrapping takes place
-		Hooks.enableAutomaticContextPropagation();
-		ArrayBlockingQueue<String> modifiedThreadLocals = new ArrayBlockingQueue<>(10);
-		ArrayBlockingQueue<String> restoredThreadLocals = new ArrayBlockingQueue<>(10);
-
-		Flux.range(0, 10)
-				.doOnNext(i -> {
-					REF1.set("i: " + i);
-				})
-				.publishOn(Schedulers.parallel())
-		        // the validation below shows that modifications to TLs are propagated
-		        // across thread boundaries via queue wrapping, so explicit control
-		        // is required from users to clean up after such modifications
-				.doOnNext(i -> modifiedThreadLocals.add(REF1.get()))
-				.contextWrite(Function.identity())
-                // the contextWrite above creates a barrier that ensures the downstream
-		        // operator sees TLs from the subscriber context
-				.doOnNext(i -> restoredThreadLocals.add(REF1.get()))
-				.contextWrite(Context.of(KEY1, "present"))
-				.blockLast();
-
-		assertThat(modifiedThreadLocals).containsExactly(
-				"i: 0", "i: 1", "i: 2", "i: 3", "i: 4",
-				"i: 5", "i: 6", "i: 7", "i: 8", "i: 9"
-		);
-		assertThat(restoredThreadLocals).containsExactly(
-				Collections.nCopies(10, "present").toArray(new String[] {})
-		);
-	}
-
-	@Test
-	void contextCapturePropagatedAutomaticallyToAllSignals() throws InterruptedException {
-		Hooks.enableAutomaticContextPropagation();
-
-		AtomicReference<String> requestTlValue = new AtomicReference<>();
-		AtomicReference<String> subscribeTlValue = new AtomicReference<>();
-		AtomicReference<String> firstNextTlValue = new AtomicReference<>();
-		AtomicReference<String> secondNextTlValue = new AtomicReference<>();
-		AtomicReference<String> cancelTlValue = new AtomicReference<>();
-
-		CountDownLatch itemDelivered = new CountDownLatch(1);
-		CountDownLatch cancelled = new CountDownLatch(1);
-
-		TestSubscriber<Integer> subscriber =
-				TestSubscriber.builder().initialRequest(1).build();
-
-		REF1.set("downstreamContext");
-
-		Flux.just(1, 2, 3)
-		    .hide()
-		    .doOnRequest(r -> requestTlValue.set(REF1.get()))
-		    .doOnNext(i -> firstNextTlValue.set(REF1.get()))
-		    .doOnSubscribe(s -> subscribeTlValue.set(REF1.get()))
-		    .doOnCancel(() -> {
-			    cancelTlValue.set(REF1.get());
-			    cancelled.countDown();
-		    })
-		    .delayElements(Duration.ofMillis(1))
-		    .contextWrite(Context.of(KEY1, "upstreamContext"))
-		    // disabling prefetching to observe cancellation
-		    .publishOn(Schedulers.parallel(), 1)
-		    .doOnNext(i -> {
-			    System.out.println(REF1.get());
-				secondNextTlValue.set(REF1.get());
-				itemDelivered.countDown();
-		    })
-		    .subscribeOn(Schedulers.boundedElastic())
-		    .contextCapture()
-		    .subscribe(subscriber);
-
-		itemDelivered.await();
-
-		subscriber.cancel();
-
-		cancelled.await();
-
-		assertThat(requestTlValue.get()).isEqualTo("upstreamContext");
-		assertThat(subscribeTlValue.get()).isEqualTo("upstreamContext");
-		assertThat(firstNextTlValue.get()).isEqualTo("upstreamContext");
-		assertThat(cancelTlValue.get()).isEqualTo("upstreamContext");
-		assertThat(secondNextTlValue.get()).isEqualTo("downstreamContext");
-	}
-
-	@Test
-	void prefetchingShouldMaintainThreadLocals() {
-		Hooks.enableAutomaticContextPropagation();
-
-		// We validate streams of items above default prefetch size
-		// (max concurrency of flatMap == Queues.SMALL_BUFFER_SIZE == 256)
-		// are able to maintain the context propagation to ThreadLocals
-		// in the presence of prefetching
-		int size = Queues.SMALL_BUFFER_SIZE * 10;
-
-		Flux<Integer> source = Flux.create(s -> {
-			for (int i = 0; i < size; i++) {
-				s.next(i);
-			}
-			s.complete();
-		});
-
-		assertThat(REF1.get()).isEqualTo("ref1_init");
-
-		ArrayBlockingQueue<String> innerThreadLocals = new ArrayBlockingQueue<>(size);
-		ArrayBlockingQueue<String> outerThreadLocals = new ArrayBlockingQueue<>(size);
-
-		source.publishOn(Schedulers.boundedElastic())
-		      .flatMap(i -> Mono.just(i)
-		                        .delayElement(Duration.ofMillis(1))
-		                        .doOnNext(j -> innerThreadLocals.add(REF1.get())))
-		      .contextWrite(ctx -> ctx.put(KEY1, "present"))
-		      .publishOn(Schedulers.parallel())
-		      .doOnNext(i -> outerThreadLocals.add(REF1.get()))
-		      .blockLast();
-
-		assertThat(innerThreadLocals).containsOnly("present").hasSize(size);
-		assertThat(outerThreadLocals).containsOnly("ref1_init").hasSize(size);
-	}
-
-	@Test
-	@Disabled("queue wrapping is removed starting from 3.5.7")
-	void queueWrapperWorksWithQueues() {
-		Hooks.enableAutomaticContextPropagation();
-		Queue<Object> queue = Queues.small()
-		                              .get();
-
-		assertThat(queue.offer("1")).isTrue();
-		assertThat(queue.poll()).isSameAs("1");
-		assertThat(queue.add("2")).isTrue();
-		assertThat(queue.remove()).isSameAs("2");
-		assertThat(queue.isEmpty()).isTrue();
-		assertThat(queue.addAll(Arrays.asList("3", "4", "5"))).isTrue();
-		assertThat(queue.peek()).isSameAs("3");
-		assertThat(queue.isEmpty()).isFalse();
-		assertThat(queue.element()).isSameAs("3");
-		assertThat(queue.size()).isEqualTo(3);
-		queue.clear();
-		assertThat(queue.offer("0")).isTrue();
-		assertThat(queue.size()).isEqualTo(1);
-
-		assertThatExceptionOfType(UnsupportedOperationException.class)
-				.isThrownBy(queue::iterator);
-		assertThatExceptionOfType(UnsupportedOperationException.class)
-				.isThrownBy(() -> queue.contains("0"));
-		assertThatExceptionOfType(UnsupportedOperationException.class)
-				.isThrownBy(queue::toArray);
-		assertThatExceptionOfType(UnsupportedOperationException.class)
-				.isThrownBy(() -> queue.toArray(new Object[] {}));
-		assertThatExceptionOfType(UnsupportedOperationException.class)
-				.isThrownBy(() -> queue.remove("0"));
-		assertThatExceptionOfType(UnsupportedOperationException.class)
-				.isThrownBy(() -> queue.containsAll(Collections.singletonList("0")));
-		assertThatExceptionOfType(UnsupportedOperationException.class)
-				.isThrownBy(() -> queue.retainAll(Collections.singletonList("5")));
-		assertThatExceptionOfType(UnsupportedOperationException.class)
-				.isThrownBy(() -> queue.removeAll(Collections.singletonList("0")));
-	}
-
-	@Test
 	void isContextPropagationAvailable() {
 		assertThat(ContextPropagationSupport.isContextPropagationAvailable()).isTrue();
 	}
@@ -359,29 +122,6 @@ class ContextPropagationTest {
 			);
 	}
 
-	@Test
-	void fluxApiUsesContextPropagationConstantFunctionWhenAutomaticPropagationEnabled() {
-		Hooks.enableAutomaticContextPropagation();
-		Flux<Integer> source = Flux.empty();
-		assertThat(source.contextCapture())
-				.isInstanceOfSatisfying(FluxContextWriteRestoringThreadLocals.class,
-						fcw -> assertThat(fcw.doOnContext)
-								.as("flux's capture function")
-								.isSameAs(ContextPropagation.WITH_GLOBAL_REGISTRY_NO_PREDICATE)
-				);
-	}
-
-	@Test
-	void monoApiUsesContextPropagationConstantFunctionWhenAutomaticPropagationEnabled() {
-		Hooks.enableAutomaticContextPropagation();
-		Mono<Integer> source = Mono.empty();
-		assertThat(source.contextCapture())
-				.isInstanceOfSatisfying(MonoContextWriteRestoringThreadLocals.class,
-						fcw -> assertThat(fcw.doOnContext)
-								.as("mono's capture function")
-								.isSameAs(ContextPropagation.WITH_GLOBAL_REGISTRY_NO_PREDICATE));
-	}
-
 	@Nested
 	class ContextCaptureFunctionTest {
 
@@ -400,203 +140,6 @@ class ContextPropagationTest {
 				.containsEntry(KEY1, "expected1")
 				.containsEntry(KEY2, "expected2")
 				.hasSize(2);
-		}
-	}
-
-	@Nested
-	class NonReactorSources {
-		@Test
-		void fluxFromPublisher() throws InterruptedException, ExecutionException {
-			ExecutorService executorService = Executors.newSingleThreadExecutor();
-			Hooks.enableAutomaticContextPropagation();
-			AtomicReference<String> value = new AtomicReference<>();
-
-			TestPublisher<String> testPublisher = TestPublisher.create();
-			Publisher<String> nonReactorPublisher = testPublisher;
-
-			Flux.from(nonReactorPublisher)
-			    .doOnNext(s -> value.set(REF1.get()))
-			    .contextWrite(Context.of(KEY1, "present"))
-			    .subscribe();
-
-			executorService
-					.submit(() -> testPublisher.emit("test").complete())
-					.get();
-
-			testPublisher.assertWasSubscribed();
-			testPublisher.assertWasNotCancelled();
-			testPublisher.assertWasRequested();
-			assertThat(value.get()).isEqualTo("present");
-
-			// validate there are no leftovers for other tasks to be attributed to
-			// previous values
-			executorService.submit(() -> value.set(REF1.get())).get();
-
-			assertThat(value.get()).isEqualTo("ref1_init");
-
-			// validate the current Thread does not have the value set either
-			assertThat(REF1.get()).isEqualTo("ref1_init");
-
-			executorService.shutdownNow();
-		}
-
-		@Test
-		void monoFromPublisher() throws InterruptedException, ExecutionException {
-			ExecutorService executorService = Executors.newSingleThreadExecutor();
-			Hooks.enableAutomaticContextPropagation();
-			AtomicReference<String> value = new AtomicReference<>();
-
-			TestPublisher<String> testPublisher = TestPublisher.create();
-			Publisher<String> nonReactorPublisher = testPublisher;
-
-			Mono.from(nonReactorPublisher)
-			    .doOnNext(s -> value.set(REF1.get()))
-			    .contextWrite(Context.of(KEY1, "present"))
-			    .subscribe();
-
-			executorService
-					.submit(() -> testPublisher.emit("test").complete())
-					.get();
-
-			testPublisher.assertWasSubscribed();
-			testPublisher.assertCancelled();
-			testPublisher.assertWasRequested();
-			assertThat(value.get()).isEqualTo("present");
-
-			// validate there are no leftovers for other tasks to be attributed to
-			// previous values
-			executorService.submit(() -> value.set(REF1.get())).get();
-
-			assertThat(value.get()).isEqualTo("ref1_init");
-
-			// validate the current Thread does not have the value set either
-			assertThat(REF1.get()).isEqualTo("ref1_init");
-
-			executorService.shutdownNow();
-		}
-
-		@Test
-		void monoFromPublisherIgnoringContract()
-				throws InterruptedException, ExecutionException {
-			ExecutorService executorService = Executors.newSingleThreadExecutor();
-			Hooks.enableAutomaticContextPropagation();
-			AtomicReference<String> value = new AtomicReference<>();
-
-			TestPublisher<String> testPublisher = TestPublisher.create();
-			Publisher<String> nonReactorPublisher = testPublisher;
-
-			Mono.fromDirect(nonReactorPublisher)
-			    .doOnNext(s -> value.set(REF1.get()))
-			    .contextWrite(Context.of(KEY1, "present"))
-			    .subscribe();
-
-			executorService
-					.submit(() -> testPublisher.emit("test").complete())
-					.get();
-
-			testPublisher.assertWasSubscribed();
-			testPublisher.assertWasNotCancelled();
-			testPublisher.assertWasRequested();
-			assertThat(value.get()).isEqualTo("present");
-
-			// validate there are no leftovers for other tasks to be attributed to
-			// previous values
-			executorService.submit(() -> value.set(REF1.get())).get();
-
-			assertThat(value.get()).isEqualTo("ref1_init");
-
-			// validate the current Thread does not have the value set either
-			assertThat(REF1.get()).isEqualTo("ref1_init");
-
-			executorService.shutdownNow();
-		}
-
-		@Test
-		void monoFromCompletionStage() throws ExecutionException, InterruptedException {
-			ExecutorService executorService = Executors.newSingleThreadExecutor();
-
-			Hooks.enableAutomaticContextPropagation();
-			CountDownLatch latch = new CountDownLatch(1);
-			AtomicReference<String> value = new AtomicReference<>();
-
-			// we need to delay delivery to ensure the completion signal is delivered
-			// on a Thread from executorService
-			CompletionStage<String> completionStage = CompletableFuture.supplyAsync(() -> {
-				try {
-					latch.await();
-				}
-				catch (InterruptedException e) {
-					// ignore
-				}
-				return "test";
-			}, executorService);
-
-			TestSubscriber<String> testSubscriber = TestSubscriber.create();
-
-			Mono.fromCompletionStage(completionStage)
-			    .doOnNext(s -> value.set(REF1.get()))
-			    .contextWrite(Context.of(KEY1, "present"))
-			    .subscribe(testSubscriber);
-
-			latch.countDown();
-			testSubscriber.block();
-
-			assertThat(value.get()).isEqualTo("present");
-
-			// validate there are no leftovers for other tasks to be attributed to
-			// previous values
-			executorService.submit(() -> value.set(REF1.get())).get();
-
-			assertThat(value.get()).isEqualTo("ref1_init");
-
-			// validate the current Thread does not have the value set either
-			assertThat(REF1.get()).isEqualTo("ref1_init");
-
-			executorService.shutdownNow();
-		}
-
-		@Test
-		void monoFromFuture() throws ExecutionException, InterruptedException {
-			ExecutorService executorService = Executors.newSingleThreadExecutor();
-
-			Hooks.enableAutomaticContextPropagation();
-			CountDownLatch latch = new CountDownLatch(1);
-			AtomicReference<String> value = new AtomicReference<>();
-
-			// we need to delay delivery to ensure the completion signal is delivered
-			// on a Thread from executorService
-			CompletableFuture<String> future = CompletableFuture.supplyAsync(() -> {
-				try {
-					latch.await();
-				}
-				catch (InterruptedException e) {
-					// ignore
-				}
-				return "test";
-			}, executorService);
-
-			TestSubscriber<String> testSubscriber = TestSubscriber.create();
-
-			Mono.fromFuture(future)
-			    .doOnNext(s -> value.set(REF1.get()))
-			    .contextWrite(Context.of(KEY1, "present"))
-			    .subscribe(testSubscriber);
-
-			latch.countDown();
-			testSubscriber.block();
-
-			assertThat(value.get()).isEqualTo("present");
-
-			// validate there are no leftovers for other tasks to be attributed to
-			// previous values
-			executorService.submit(() -> value.set(REF1.get())).get();
-
-			assertThat(value.get()).isEqualTo("ref1_init");
-
-			// validate the current Thread does not have the value set either
-			assertThat(REF1.get()).isEqualTo("ref1_init");
-
-			executorService.shutdownNow();
 		}
 	}
 
@@ -953,127 +496,6 @@ class ContextPropagationTest {
 				      .as("subFused.handler")
 				      .isNotSameAs(originalHandler);
 			});
-		}
-	}
-
-	@Nested
-	class BlockingOperatorsAutoCapture {
-
-		@Test
-		void monoBlock() {
-			Hooks.enableAutomaticContextPropagation();
-
-			AtomicReference<String> value = new AtomicReference<>();
-
-			REF1.set("present");
-
-			Mono.just("test")
-			    // Introduce an artificial barrier to clear ThreadLocals if no Context
-			    // is defined in the downstream chain. If block does the job well,
-			    // it should have captured the existing ThreadLocal into the Context.
-			    .contextWrite(Context.empty())
-			    .doOnNext(ignored -> value.set(REF1.get()))
-			    .block();
-
-			// First, assert the existing ThreadLocal was not cleared.
-			assertThat(REF1.get()).isEqualTo("present");
-
-			// Now let's find out that it was automatically transferred.
-			assertThat(value.get()).isEqualTo("present");
-		}
-
-		@Test
-		void monoBlockOptional() {
-			Hooks.enableAutomaticContextPropagation();
-
-			AtomicReference<String> value = new AtomicReference<>();
-
-			REF1.set("present");
-
-			Mono.empty()
-			    // Introduce an artificial barrier to clear ThreadLocals if no Context
-			    // is defined in the downstream chain. If block does the job well,
-			    // it should have captured the existing ThreadLocal into the Context.
-			    .contextWrite(Context.empty())
-			    .doOnTerminate(() -> value.set(REF1.get()))
-			    .blockOptional();
-
-			// First, assert the existing ThreadLocal was not cleared.
-			assertThat(REF1.get()).isEqualTo("present");
-
-			// Now let's find out that it was automatically transferred.
-			assertThat(value.get()).isEqualTo("present");
-		}
-
-		@Test
-		void fluxBlockFirst() {
-			Hooks.enableAutomaticContextPropagation();
-
-			AtomicReference<String> value = new AtomicReference<>();
-
-			REF1.set("present");
-
-			Flux.range(0, 10)
-			    // Introduce an artificial barrier to clear ThreadLocals if no Context
-			    // is defined in the downstream chain. If block does the job well,
-			    // it should have captured the existing ThreadLocal into the Context.
-			    .contextWrite(Context.empty())
-			    .doOnNext(ignored -> value.set(REF1.get()))
-			    .blockFirst();
-
-			// First, assert the existing ThreadLocal was not cleared.
-			assertThat(REF1.get()).isEqualTo("present");
-
-			// Now let's find out that it was automatically transferred.
-			assertThat(value.get()).isEqualTo("present");
-		}
-
-		@Test
-		void fluxBlockLast() {
-			Hooks.enableAutomaticContextPropagation();
-
-			AtomicReference<String> value = new AtomicReference<>();
-
-			REF1.set("present");
-
-			Flux.range(0, 10)
-			    // Introduce an artificial barrier to clear ThreadLocals if no Context
-			    // is defined in the downstream chain. If block does the job well,
-			    // it should have captured the existing ThreadLocal into the Context.
-			    .contextWrite(Context.empty())
-			    .doOnTerminate(() -> value.set(REF1.get()))
-			    .blockLast();
-
-			// First, assert the existing ThreadLocal was not cleared.
-			assertThat(REF1.get()).isEqualTo("present");
-
-			// Now let's find out that it was automatically transferred.
-			assertThat(value.get()).isEqualTo("present");
-		}
-
-		@Test
-		void fluxToIterable() {
-			Hooks.enableAutomaticContextPropagation();
-
-			AtomicReference<String> value = new AtomicReference<>();
-
-			REF1.set("present");
-
-			Iterable<Integer> integers = Flux.range(0, 10)
-			                                 // Introduce an artificial barrier to clear ThreadLocals if no Context
-			                                 // is defined in the downstream chain. If block does the job well,
-			                                 // it should have captured the existing ThreadLocal into the Context.
-			                                 .contextWrite(Context.empty())
-			                                 .doOnTerminate(() -> value.set(REF1.get()))
-			                                 .toIterable();
-
-			assertThat(integers).hasSize(10);
-
-			// First, assert the existing ThreadLocal was not cleared.
-			assertThat(REF1.get()).isEqualTo("present");
-
-			// Now let's find out that it was automatically transferred.
-			assertThat(value.get()).isEqualTo("present");
 		}
 	}
 }
